pyproject.toml
[tool.poetry]
name = "restful-api-gateway"
version = "0.1.0"
description = ""
authors = ["Blake Enyart <blake.enyart@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.62.0"
"aws-cdk.aws-lambda" = "^1.87.0"
"aws-cdk.aws-apigateway" = "^1.87.0"

[tool.poetry.dev-dependencies]
invoke = "^1.5.0"
black = "^20.8b1"
klaxon = "^1.1.1"
uvicorn = "^0.13.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
